+++ packages/web/components/common/Icon/constants.ts
@@ -97,8 +97,6 @@ export const iconPaths = {
   'core/app/variable/textarea': () => import('./icons/core/app/variable/textarea.svg'),
   'core/chat/QGFill': () => import('./icons/core/chat/QGFill.svg'),
   'core/chat/cancelSpeak': () => import('./icons/core/chat/cancelSpeak.svg'),
-  'core/chat/chevronSelector': () => import('./icons/core/chat/chevronSelector.svg'),
-  'core/chat/sideLine': () => import('./icons/core/chat/sideLine.svg'),
   'core/chat/chatFill': () => import('./icons/core/chat/chatFill.svg'),
   'core/chat/chatLight': () => import('./icons/core/chat/chatLight.svg'),
   'core/chat/chatModelTag': () => import('./icons/core/chat/chatModelTag.svg'),
