+++ projects/app/src/components/core/chat/ChatContainer/ChatBox/components/ChatItem.tsx
@@ -1,5 +1,5 @@
 import { Box, BoxProps, Card, Flex } from '@chakra-ui/react';
-import React, { useMemo } from 'react';
+import React, { useMemo, useTransition } from 'react';
 import ChatController, { type ChatControllerProps } from './ChatController';
 import ChatAvatar from './ChatAvatar';
 import { MessageCardStyle } from '../constants';
@@ -11,7 +11,10 @@ import FilesBlock from './FilesBox';
 import { ChatBoxContext } from '../Provider';
 import { useContextSelector } from 'use-context-selector';
 import AIResponseBox from '../../../components/AIResponseBox';
-
+import { useCopyData } from '@/web/common/hooks/useCopyData';
+import MyIcon from '@fastgpt/web/components/common/Icon';
+import MyTooltip from '@fastgpt/web/components/common/MyTooltip';
+import { useTranslation } from 'next-i18next';
 const colorMap = {
   [ChatStatusEnum.loading]: {
     bg: 'myGray.100',
@@ -62,9 +65,11 @@ const ChatItem = ({
           bg: 'myGray.50'
         };
 
+  const { t } = useTranslation();
   const isChatting = useContextSelector(ChatBoxContext, (v) => v.isChatting);
   const { chat } = chatControllerProps;
-
+  const { copyData } = useCopyData();
+  const chatText = useMemo(() => formatChatValue2InputType(chat.value).text || '', [chat.value]);
   const ContentCard = useMemo(() => {
     if (type === 'Human') {
       const { text, files = [] } = formatChatValue2InputType(chat.value);
@@ -148,6 +153,28 @@ const ChatItem = ({
         >
           {ContentCard}
           {children}
+          {/* 对话框底部的复制按钮 */}
+          {type == ChatRoleEnum.AI && (!isChatting || (isChatting && !isLastChild)) && (
+            <Box
+              position={'absolute'}
+              bottom={0}
+              right={[0, -2]}
+              color={'myGray.400'}
+              transform={'translateX(100%)'}
+            >
+              <MyTooltip label={t('common.Copy')}>
+                <MyIcon
+                  w={'14px'}
+                  cursor="pointer"
+                  p="5px"
+                  bg="white"
+                  name={'copy'}
+                  _hover={{ color: 'primary.600' }}
+                  onClick={() => copyData(chatText)}
+                />
+              </MyTooltip>
+            </Box>
+          )}
         </Card>
       </Box>
     </>
